Fix average precision

+++ src/org/zeroxlab/benchmark/CaseCanvas.java
@@ -62,7 +62,7 @@ public class CaseCanvas extends Case{
         }
 
         String result = "";
-        long total = 0;
+        float total = 0;
         int length = mResult.length;
 
         for (int i = 0; i < length; i++) {
